Build libecc for NetBSD (as a sanity check for now).
git-svn-id: http://ellcc.org/svn/ellcc/trunk@2348 82857169-6e38-49d7-aaa8-07fb558f12fc

+++ libecc/src/syscalls/lseek.c
@@ -4,5 +4,5 @@
 
 off_t lseek(int fd, off_t offset, int whence)
 {
-    return INLINE_SYSCALL(lseek, 3, fd, offset, whence);
+    return INLINE_SYSCALL(lseek, 3, fd, (long)offset, whence);  // RICH: Netbsd is 64 bit.
 }
